requirements.txt
pyttsx3~=2.90
wikipedia~=1.4.0
PyAutoGUI~=0.9.54
SpeechRecognition~=3.10.4
pyjokes==0.6.0
pywinauto~=0.6.8
PyGetWindow~=0.0.9
pyperclip~=1.9.0
Pillow~=11.2.1
pystray~=0.19.5
python-dotenv~=1.0.1
openai~=1.79.0
yt-dlp~=2025.3.31
